+++ monkey/common/agent_event_serializers/agent_event_serializer_registry.py
@@ -1,15 +1,15 @@
 from typing import Type, Union
 
-from common.event_serializers import IEventSerializer
-from common.events import AbstractAgentEvent
+from common.agent_event_serializers import IAgentEventSerializer
+from common.agent_events import AbstractAgentEvent
 
 
-class EventSerializerRegistry:
+class AgentEventSerializerRegistry:
     """
     Registry for event serializers using event class.
 
     Example:
-        event_serializer_registry = EventSerializerRegistry()
+        event_serializer_registry = AgentEventSerializerRegistry()
         event_serializer_registry[MyEvent] = MyEventSerializer()
 
         my_event_dict = {"type": "MyEvent", "data": "123"}
@@ -22,18 +22,20 @@ class EventSerializerRegistry:
         self._registry = {}
 
     def __setitem__(
-        self, event_class: Type[AbstractAgentEvent], event_serializer: IEventSerializer
+        self, event_class: Type[AbstractAgentEvent], event_serializer: IAgentEventSerializer
     ):
         if not issubclass(event_class, AbstractAgentEvent):
             raise TypeError(f"Event class must be of type: {AbstractAgentEvent.__name__}")
 
-        if not isinstance(event_serializer, IEventSerializer):
-            raise TypeError(f"Event serializer must be of type: {IEventSerializer.__name__}")
+        if not isinstance(event_serializer, IAgentEventSerializer):
+            raise TypeError(f"Event serializer must be of type: {IAgentEventSerializer.__name__}")
 
         self._registry[event_class] = event_serializer
         self._registry[event_class.__name__] = event_serializer
 
-    def __getitem__(self, event_class: Union[str, Type[AbstractAgentEvent]]) -> IEventSerializer:
+    def __getitem__(
+        self, event_class: Union[str, Type[AbstractAgentEvent]]
+    ) -> IAgentEventSerializer:
         if not (isinstance(event_class, str) or issubclass(event_class, AbstractAgentEvent)):
             raise TypeError(
                 f"Registry get key {event_class} must be of type: {AbstractAgentEvent.__name__} or "