+++ monkey/common/event_serializers/register.py
@@ -1,9 +1,0 @@
-from common.events import CredentialsStolenEvent
-
-from . import EventSerializerRegistry, PydanticEventSerializer
-
-
-def register_common_agent_event_serializers(event_serializer_registry: EventSerializerRegistry):
-    event_serializer_registry[CredentialsStolenEvent] = PydanticEventSerializer(
-        CredentialsStolenEvent
-    )