+++ monkey/monkey_island/cc/app.py
@@ -13,9 +13,9 @@ from monkey_island.cc.database import database, mongo
 from monkey_island.cc.resources import (
     AgentBinaries,
     AgentConfiguration,
+    AgentEvents,
     Agents,
     ClearSimulationData,
-    Events,
     IPAddresses,
     IslandLog,
     PBAFileDownload,
@@ -187,7 +187,7 @@ def init_restful_endpoints(api: FlaskDIWrapper):
     api.add_resource(IslandLog)
     api.add_resource(IPAddresses)
 
-    api.add_resource(Events)
+    api.add_resource(AgentEvents)
 
     # API Spec: These two should be the same resource, GET for download and POST for upload
     api.add_resource(PBAFileDownload)