+++ monkey/monkey_island/cc/repository/__init__.py
@@ -10,7 +10,7 @@ from .i_user_repository import IUserRepository
 from .i_machine_repository import IMachineRepository
 from .i_agent_repository import IAgentRepository
 from .i_node_repository import INodeRepository
-from .i_event_repository import IEventRepository
+from .i_agent_event_repository import IAgentEventRepository
 
 
 from .local_storage_file_repository import LocalStorageFileRepository