+++ monkey/monkey_island/cc/repository/i_agent_event_repository.py
@@ -1,13 +1,13 @@
 from abc import ABC, abstractmethod
 from typing import Sequence, Type, TypeVar
 
-from common.events import AbstractAgentEvent
+from common.agent_events import AbstractAgentEvent
 from common.types import AgentID
 
 T = TypeVar("T", bound=AbstractAgentEvent)
 
 
-class IEventRepository(ABC):
+class IAgentEventRepository(ABC):
     """A repository used to store and retrieve event objects"""
 
     @abstractmethod