+++ monkey/monkey_island/cc/repository/mongo_event_repository.py
@@ -2,19 +2,21 @@ from typing import Any, Dict, MutableMapping, Sequence, Type
 
 from pymongo import MongoClient
 
-from common.event_serializers import EVENT_TYPE_FIELD, EventSerializerRegistry
-from common.events import AbstractAgentEvent
+from common.agent_event_serializers import EVENT_TYPE_FIELD, AgentEventSerializerRegistry
+from common.agent_events import AbstractAgentEvent
 from common.types import AgentID
-from monkey_island.cc.repository import IEventRepository
+from monkey_island.cc.repository import IAgentEventRepository
 
 from . import RemovalError, RetrievalError, StorageError
 from .consts import MONGO_OBJECT_ID_KEY
 
 
-class MongoEventRepository(IEventRepository):
+class MongoEventRepository(IAgentEventRepository):
     """A repository for storing and retrieving events in MongoDB"""
 
-    def __init__(self, mongo_client: MongoClient, serializer_registry: EventSerializerRegistry):
+    def __init__(
+        self, mongo_client: MongoClient, serializer_registry: AgentEventSerializerRegistry
+    ):
         self._events_collection = mongo_client.monkey_island.events
         self._serializers = serializer_registry