+++ monkey/monkey_island/cc/server_setup.py
@@ -23,8 +23,8 @@ if str(MONKEY_ISLAND_DIR_BASE_PATH) not in sys.path:
     sys.path.insert(0, MONKEY_ISLAND_DIR_BASE_PATH)
 
 from common import DIContainer  # noqa: E402
-from common.event_serializers import (  # noqa: E402
-    EventSerializerRegistry,
+from common.agent_event_serializers import (  # noqa: E402
+    AgentEventSerializerRegistry,
     register_common_agent_event_serializers,
 )
 from common.network.network_utils import get_my_ip_addresses  # noqa: E402
@@ -140,10 +140,10 @@ def _initialize_di_container(
 
 
 def _setup_agent_event_serializers(container: DIContainer):
-    agent_event_serializer_registry = EventSerializerRegistry()
+    agent_event_serializer_registry = AgentEventSerializerRegistry()
     register_common_agent_event_serializers(agent_event_serializer_registry)
 
-    container.register_instance(EventSerializerRegistry, agent_event_serializer_registry)
+    container.register_instance(AgentEventSerializerRegistry, agent_event_serializer_registry)
 
 
 def _initialize_mongodb_connection(start_mongodb: bool, data_dir: Path):