+++ monkey/tests/unit_tests/common/event_queue/test_pypubsub_agent_event_queue.py
@@ -5,8 +5,8 @@ from uuid import UUID
 import pytest
 from pubsub.core import Publisher
 
+from common.agent_events import AbstractAgentEvent
 from common.event_queue import AgentEventSubscriber, IAgentEventQueue, PyPubSubAgentEventQueue
-from common.events import AbstractAgentEvent
 
 EVENT_TAG_1 = "event tag 1"
 EVENT_TAG_2 = "event tag 2"