+++ CHANGELOG.md
@@ -20,7 +20,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 - `/api/clear-simulation-data` endpoint. #2036
 - `/api/registration-status` endpoint. #2149
 - authentication to `/api/island/version`. #2109
-- `/api/events` endpoint. #2155
+- `/api/agent-events` endpoint. #2155, #2300
 - The ability to customize the file extension used by ransomware when
   encrypting files. #1242
 - `/api/agents` endpoint.